Modularize jobs
